backend/requirements.txt
tensorflow-cpu>=2.10.0
tensorflow-cpu>=2.10.0
keras>=2.10.0
flask>=2.0.0
flask-cors>=3.0.10
opencv-python>=4.5.0
numpy>=1.20.0
pillow>=8.0.0
pandas>=1.2.0
scikit-learn>=0.24.0
gunicorn>=20.1.0
